backend/requirements.txt
fastapi[standard]>=0.115,<1
uvicorn[standard]>=0.32,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
pymupdf>=1.24,<2
chromadb>=0.5,<1
rank-bm25>=0.2.2,<1
